vision_models/requirements.txt
torchvision==0.13.0
timm
mmdet
boto3
google-cloud-vision
clip-anytorch
azure-cognitiveservices-vision-computervision